+++ src/idc/imgaug/filter/__init__.py
@@ -7,6 +7,7 @@ from ._base_filter import BaseFilter
 from ._base_image_augmentation import BaseImageAugmentation
 from ._change_grayscale import ChangeGrayscale
 from ._clip_grayscale import ClipGrayscale
+from ._code_detector import CodeDetector
 from ._crop import Crop
 from ._crop_to_label import CropToLabel
 from ._enhance import Enhance, ENHANCEMENTS, ENHANCEMENT_CONTRAST, ENHANCEMENT_COLOR, ENHANCEMENT_SHARPNESS, ENHANCEMENT_BRIGHTNESS