+++ CHANGES.rst
@@ -19,6 +19,7 @@ Changelog
   meta-data as well
 - added `merge-segmentation-layers` filter to merging the layer of image segmentation item in storage with
   the one currently passing through filter
+- added `code-detector` filter for detecting 1-dimensional barcodes and QR codes
 
 
 0.1.0 (2025-10-31)